Prompt 405B
Input 
The first line contains a single integer n (1 ≤ n ≤ 3000), the number of the dominoes in the line. The next line contains a character string s of length n. The i-th character of the string si is equal to

"L", if the i-th domino has been pushed to the left;
"R", if the i-th domino has been pushed to the right;
".", if the i-th domino has not been pushed.
It is guaranteed that if si = sj = "L" and i < j, then there exists such k that i < k < j and sk = "R"; if si = sj = "R" and i < j, then there exists such k that i < k < j and sk = "L".

Output
Output a single integer, the number of the dominoes that remain vertical at the end of the process.

Examples
input
14
.L.R...LR..L..
output
4